requirements.txt
django
djangorestframework
django-cors-headers
pyscopg2 
